Move all react components to ui folder
